Rename expense summary card file
