+++ lib/carmanager_ui.dart
@@ -28,7 +28,7 @@ export 'package:carmanager_ui/src/components/dialogs/error/error_dialog_route.da
 export 'package:carmanager_ui/src/components/dialogs/loading/loading_dialog_data.dart';
 export 'package:carmanager_ui/src/components/dialogs/loading/loading_dialog_route.dart';
 export 'package:carmanager_ui/src/components/dropdown/cm_dropdown.dart';
-export 'package:carmanager_ui/src/components/expense_summary_card.dart';
+export 'package:carmanager_ui/src/components/cm_expense_summary_card.dart';
 export 'package:carmanager_ui/src/components/loader/loader.dart';
 export 'package:carmanager_ui/src/components/report_tile/report_tile.dart';
 export 'package:carmanager_ui/src/components/text_field/cm_multiline_text_field.dart';
